- Overloaded new to force 16 byte alignment. - Added configure check for 16 byte alignment. - Added more verbose check. - Added alignment attribute for GCC. - Corrected constant loading of SSE version of b3Color.
And finally:
- Corrected distance check inside triangle field
  intersectin computation from b3_f32 to more
  accurate b3_f64. This error forced a seg fault due to
  a errornous index computation into the triangle field.


git-svn-id: svn+ssh://svn.morknet.de/usr/local/subversion/blz3/trunk@1475 c0f852eb-cc16-0410-a65c-258990c0cc53

+++ src/system/b3Exception.cc
@@ -105,7 +105,7 @@ const char *b3ExceptionBase::what() const throw()
 	char buffer[128];
 
 	b3GetErrorMsg();
-	snprintf(buffer, sizeof(buffer), " file: %-32.32s line: %5d",m_FileName, m_LineNo);
+	snprintf(buffer, sizeof(buffer), " file: %-32.32s line: %5d",m_FileName, static_cast<int>(m_LineNo));
 	strncat(m_LocalMessageBuffer, buffer, sizeof(m_LocalMessageBuffer));
 
 	return m_LocalMessageBuffer;
